+++ app/src/main/java/com/openclassrooms/realestatemanager/ui/listview/ListViewFragment.java
@@ -3,21 +3,14 @@ package com.openclassrooms.realestatemanager.ui.listview;
 import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.ViewModelProvider;
-import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -29,7 +22,6 @@ import com.openclassrooms.realestatemanager.ui.addedit.AddEditGeneralFragment;
 import com.openclassrooms.realestatemanager.ui.details.DetailsFragment;
 import com.openclassrooms.realestatemanager.viewmodels.ListViewModel;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -67,7 +59,7 @@ public class ListViewFragment extends Fragment implements ListViewAdapter.OnItem
     }
 
     @Override
-    public void onItemClick(int id) {
+    public void onItemClick(long id) {
         Fragment fragment = DetailsFragment.newInstance(id);
         FragmentTransaction transaction = getActivity().getSupportFragmentManager().beginTransaction();
         transaction.replace(R.id.nav_host_fragment, fragment, "details_fragment");
